Prop出力用コードで型の間違いを修正

+++ eglibrary-android-app-framework/src/main/gen-prop/com/eaglesakura/android/framework/db/BasicSettings.java
@@ -16,8 +16,8 @@ public class BasicSettings extends com.eaglesakura.android.db.BasePropertiesData
     }
     public void setGcmToken(String set){ setProperty("BasicSettings.gcmToken", set); }
     public String getGcmToken(){ return getStringProperty("BasicSettings.gcmToken"); }
-    public void setLastBootedAppVersionCode(long set){ setProperty("BasicSettings.lastBootedAppVersionCode", set); }
-    public long getLastBootedAppVersionCode(){ return getLongProperty("BasicSettings.lastBootedAppVersionCode"); }
+    public void setLastBootedAppVersionCode(int set){ setProperty("BasicSettings.lastBootedAppVersionCode", set); }
+    public long getLastBootedAppVersionCode(){ return getIntProperty("BasicSettings.lastBootedAppVersionCode"); }
     public void setLastBootedAppVersionName(String set){ setProperty("BasicSettings.lastBootedAppVersionName", set); }
     public String getLastBootedAppVersionName(){ return getStringProperty("BasicSettings.lastBootedAppVersionName"); }
